refactor: consolidate useSandboxQueries functionality
- Moved useSandboxQueries logic from useQueries to a new file for better organization.
- Updated index.ts to export the new useSandboxQueries hook.
- Enhanced useSandboxQueries to utilize useEffect for handling route changes.
