+++ src/hooks/index.ts
@@ -3,5 +3,6 @@ export { default as useFormValidation } from './useFormValidation';
 export { default as useOAuth } from './useOAuth';
 export { default as useOAuthRedirect } from './useOAuthRedirect';
 export { default as useQueryParams } from './useQueryParams';
+export { default as useQueries, default as useSandboxQueries } from './useSandboxQueries';
 export { default as useServerStats } from './useServerStats';
 export { default as useToast } from './useToast';
